[K1P-209] feat : 정기, 단건 장바구니, 바로 주문 기능 수정
