+++ frontend/sellution-app/src/shopping/business/onetimeDelivery/useDetail.js
@@ -3,6 +3,7 @@ import { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import useClientName from '../layout/useClientName';
 import useOrderListStore from './../../store/stores/useOrderListStore';
+import useOnetimeCartStore from '../../store/stores/useOnetimeCartStore';
 
 const useDetail = () => {
   const navigate = useNavigate();
@@ -32,10 +33,34 @@ const useDetail = () => {
     setCurrentImageIndex(index);
   };
 
+  const { onetimeCart, updateOnetimeCart } = useOnetimeCartStore();
+  const addToOnetimeCart = () => {
+    //stock대신 quantity를 가진 newItem 객체 만들어서, 장바구니에 넣기
+    if (itemCountToAdd > 0) {
+      const newItem = {
+        id: productToShow.productId,
+        productId: productToShow.productId,
+        quantity: itemCountToAdd,
+        name: productToShow.name,
+        cost: productToShow.cost,
+        discountRate: productToShow.discountRate,
+        discountedPrice: productToShow.discountedPrice,
+        stock: productToShow.stock,
+      };
+      updateOnetimeCart(newItem);
+      console.log('단건배송 장바구니 목록: ', onetimeCart);
+      setItemCountToAdd(0); //장바구니에 넣어준 다음에 수량 0으로 다시 변경 >> 장바구니 버튼 누르면 0으로 초기화 되는 것
+      //modal창 띄우기
+      // setIsDetailPageModalVisible(true);
+      navigate(`/shopping/${clientName}/onetime`);
+    }
+  };
+
   const handleDirectOrder = () => {
     if (itemCountToAdd > 0) {
       const newItem = {
         id: productToShow.productId,
+        productId: productToShow.productId,
         quantity: itemCountToAdd,
         name: productToShow.name,
         cost: productToShow.cost,
@@ -79,6 +104,7 @@ const useDetail = () => {
     toggleDetailOption,
     handleDirectOrder,
     handleSlideChange,
+    addToOnetimeCart,
   };
 };
 export default useDetail;
