+++ frontend/sellution-app/src/shopping/component/subscriptionDelivery/DetailComponent.jsx
@@ -4,32 +4,9 @@ import MenuHeaderNav from './../../layout/MenuHeaderNav';
 import DetailOption from '../../layout/partials/DetailOption';
 import OneButtonFooterLayout from './../../layout/OneButtonFooterLayout';
 import TwoButtonFooterLayout from '../../layout/TwoButtonFooterLayout';
-import { useNavigate } from 'react-router-dom';
-import useClientName from '../../business/layout/useClientName';
-import useSubscriptionCartStore from './../../store/stores/useSubscriptionCartStore';
 import DetailInformationLayout from '../../layout/DetailInformationLayout';
 const DetailComponent = () => {
-  const navigate = useNavigate();
-  const { clientName } = useClientName();
-  const { subscriptionCart, updateSubscriptionCart } = useSubscriptionCartStore();
-  const addToSubscriptionCart = () => {
-    if (itemCountToAdd > 0) {
-      const newItem = {
-        id: productToShow.productId,
-        quantity: itemCountToAdd,
-        name: productToShow.name,
-        cost: productToShow.cost,
-        discountRate: productToShow.discountRate,
-        discountedPrice: productToShow.discountedPrice,
-        stock: productToShow.stock,
-      };
-      updateSubscriptionCart(newItem);
-      console.log('정기배송 장바구니 목록: ', subscriptionCart);
-      setItemCountToAdd(0);
-      //setIsDetailPageModalVisible(true);
-      navigate(`/shopping/${clientName}/subscription/cart`);
-    }
-  };
+
   const {
     activeSlide,
     setActiveSlide,
@@ -44,6 +21,7 @@ const DetailComponent = () => {
     toggleDetailOption,
     handleDirectOrder,
     handleSlideChange,
+    addToSubscriptionCart,
   } = useDetail();
 
   if (isLoading) {
