+++ frontend/sellution-app/src/shopping/store/stores/useOrderListStore.js
@@ -4,8 +4,8 @@ const useOrderListStore = create((set) => ({
   orderList: [],
 
   updateOrderList: (selectedItemsIndexList, cart) => {
-    //selectedItems에는 체크한 product의 id들만 있음
-    const selectedProducts = cart.filter((item) => selectedItemsIndexList.includes(item.id));
+    // selectedItemsIndexList에는 체크한 product의 productId들만 있음
+    const selectedProducts = cart.filter((item) => selectedItemsIndexList.includes(item.productId));
     set((state) => ({
       ...state,
       orderList: selectedProducts,
